% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVS.R
\name{GVS}
\alias{GVS}
\title{Check the validity of coordinates}
\usage{
GVS(occurrence_dataframe, ...)
}
\arguments{
\item{occurrence_dataframe}{A properly formatted dataframe, see https://bien.nceas.ucsb.edu/bien/tools/nsr/batch-mode/}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
Dataframe containing NSR results.
}
\description{
GVS returns information on coordinate validity.
}
\examples{
\dontrun{

results <- GVS(occurrence_dataframe = gvs_testfile)
}
occurrence_dataframe <-data
}
